==> 3_dimensionality_reduction.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# tSNE comparison 

```{r}
suppressPackageStartupMessages(library(tidyverse))
suppressPackageStartupMessages(library(flowCore))
suppressPackageStartupMessages(library(FlowSOM))
suppressPackageStartupMessages(library(CytoNorm))
suppressPackageStartupMessages(library(flowWorkspace))
suppressPackageStartupMessages(library(CytoML))
```

```{r}
dir <- "D:/Data/Ongoing/Stanford/data/Gates_ExtraControls"
files <- list.files(dir, pattern="Gates_P.*fcs$")
df <- data.frame(File = files,
                 Plate = as.factor(str_match(files, "(PTLG[0-9]*)")[, 2]),
                 Stim = as.factor(str_match(files, "[0-9]_(.*)_Control")[, 2]),
                 Volunteer = as.factor(str_match(files, "_([12]).")[, 2]),
                 stringsAsFactors = FALSE)
rownames(df) <- df$File
df <- df %>% dplyr::filter(Stim == "Unstim" & Volunteer == "2")
df
```

```{r}
set.seed(1)

data <- list("Original"  = AggregateFlowFrames(file.path(dir, df$File), 10000),
             "Normalized" = 
               AggregateFlowFrames(
                 file.path("results_CytoNorm/data/Cluster_QuantileNormLimited_Unstim_and_IFNa_LPS_1", 
                           paste0("Norm_", df$File)),
                 10000))
```


```{r}
cols_of_interest <- c(48, 46, 43, 45, 20, 16, 21, 19, 22, 50, 47, 40, 44, 33, 17,
                      11, 18, 51, 14, 23, 32, 10, 49, 27, 24, 31, 42, 37, 39, 34, 
                      41, 26, 30, 28, 29, 25, 35)
```

```{r}
data <- lapply(data, function(d){transform(d, transformList(colnames(d)[cols_of_interest],
                                                                      cytofTransform))})
```

```{r}
set.seed(1)

dim_red <- list()
for(d in names(data)){
  dim_red[[d]] <- Rtsne::Rtsne(data[[d]]@exprs[, cols_of_interest],
                                perplexity = 30)
}
```

```{r fig.width = 15}
set.seed(1)
rand_order <- sample(nrow(data$Original))

plots <- list()
for(d in names(data)){
  plots[[d]] <- ggplot(data.frame(x = dim_red[[d]]$Y[rand_order,1], 
                                  y = dim_red[[d]]$Y[rand_order,2],
                                  File = factor(data[[d]]@exprs[rand_order,"File"]))) +
    geom_point(aes(x, y, color = File), 
               size = 0.5) + 
    xlab("tSNE 1") + 
    ylab("tSNE 2") +
    theme_minimal() +
    ggtitle(d) +
    theme(legend.position = "none")
}

p <- do.call(gridExtra::grid.arrange, c(plots, list(ncol = 2)))

ggsave("results_CytoNorm/tSNE_before_after.pdf",
       p,
       width = 10, height = 5)

```
